Remove setup_scripts/setup_ssh_ecryptfs.sh
Sadly, ecryptfs is unmaintained. Last commit is from 2017 and there are
critical bugs;
https://bazaar.launchpad.net/~ecryptfs/ecryptfs/trunk/files

I really liked the mount on login of ecryptfs. The best alternative
seems to be https://github.com/google/fscrypt/ which uses ext4 native
encryption. It seems to require linux kernel 5.4 to fully work. It's
fine because that's what is included in Ubuntu 20.04. It's fairly recent
so I think I'll wait for folks to loose their home dir first before
trying. :)

The next best fit is probably this one but I don't think it's necessary.
https://github.com/rfjakob/gocryptfs